refactor(FileUploaderSkeleton): deduplicate FileUploaderSkeleton
