+++ src/components/FileUploader/FileUploader-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import FileUploader, { FileUploaderButton } from '../FileUploader';
-import FileUploaderSkeleton from '../FileUploader/FileUploader.Skeleton';
+import FileUploaderSkeleton from '../FileUploaderSkeleton/FileUploader.Skeleton';
 import FileUploadStatus from '../FileUploadStatus/FileUploadStatus';
 import { mount, shallow } from 'enzyme';
