+++ src/components/FileUploader/index.js
@@ -1,3 +1,3 @@
-export * from './FileUploader.Skeleton';
+export * from '../FileUploaderSkeleton/FileUploader.Skeleton';
 export * from './FileUploader';
 export default from './FileUploader';
