+++ src/components/FileUploaderV2/index.js
@@ -1,3 +1,3 @@
-export * from './FileUploaderV2.Skeleton';
+export * from '../FileUploaderSkeleton/FileUploader.Skeleton';
 export * from './FileUploaderV2';
 export default from './FileUploaderV2';
